backend/pyproject.toml
[project]
name = "customer-success-fte"
version = "1.0.0"
description = "Customer Success Digital FTE - 24/7 AI employee for multi-channel customer support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # Database
    "sqlmodel>=0.0.21",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    # Settings & Validation
    "pydantic>=2.8.0",
    "pydantic-settings>=2.3.0",
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # Forms & Uploads
    "python-multipart>=0.0.9",
    # AI & Context
    "openai>=1.35.0",
    # Kafka
    "kafka-python-ng>=2.2.0",
    # HTTP Client
    "httpx>=0.27.0",
    "pgvector>=0.4.2",
    "python-dotenv>=1.2.1",
    "email-validator>=2.3.0",
    # OpenAI Agents SDK (specific version to avoid import conflicts)
    "openai-agents>=0.10.0",
    "google-generativeai>=0.8.6",
    # LLM Provider Abstraction
    "litellm>=1.50.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
